+++ src/main/java/com/example/network/ServerMessage.java
@@ -1,14 +1,13 @@
 package com.example.network;
 
-import java.math.BigInteger;
 import java.util.LinkedHashMap;
 import java.util.List;
 
 public sealed interface ServerMessage {
-    record GameStartMessage(int clientId, long startFrame, BigInteger seed) implements ServerMessage {
+    record GameStartMessage(int clientId, long startFrame, long seed) implements ServerMessage {
     }
 
-    record StateBroadcastMessage(BigInteger messageFrame, int numClients,
+    record StateBroadcastMessage(long messageFrame, int numClients,
                                  LinkedHashMap<Integer, List<Integer>> clientStates) implements ServerMessage {
     }
